RequireJS: Use RequireJS CommonJS-package support
(package, not module - package = directory tree from NPM, which may
include RJS-compatible AMD modules, module = something that can be
imported using the CommonJS module loader).

Many packages on NPM and Bower do not consist only of a single "main"
Javascript file that can be imported to get full access to the library.
Rather, they *may* have a main JS (e.g. at path foo/bar.js), and will as
well have other files in e.g. lib/SomeLibraryModule.js. See for example
the react, react-lib, and react-router packages.

RequireJS can work with these, but to do so it needs a distinction to be
drawn between the root path of the library ("location") and the path to
any main JS file relative to that root.

The new configs look (when pretty-printed) like:

```
{
    "packages": [
        {
            "name": "objectpath",
            "location": ["http://cdn.jsdelivr.net/webjars/objectpath/1.1.0/", "/webjars/objectpath/1.1.0/"],
            "main": "lib/ObjectPath"
        }
    ]
}
```

Configs generated from Classic WebJars are unchanged.

Tests are updated in the next commit.